Split Core from extio
Get ready for libsddc development
