+++ CMakeLists.txt
@@ -1,0 +1,5 @@
+cmake_minimum_required(VERSION 3.13)
+project(SDDC)
+
+add_subdirectory(CORE)
+add_subdirectory(ExtIO_sddc)
