+++ Core/arch/win32/CyAPI/CyAPI.cpp
