+++ Core/arch/win32/CyAPI/CyUSB30_def.h
