+++ Core/arch/win32/CyAPI/cyioctl.h
