+++ Core/arch/win32/CyAPI/usb100.h
