+++ Core/arch/win32/FFTW3/COPYING
