+++ Core/framework.h
